requirements.txt
python-dotenv
langchain
langchain-mcp-adapters
mcp
langgraph
tavily
langchain-community
langsmith